requirements.txt
absl-py==2.1.0
alphafold-colabfold==2.3.7
biopython @ file:///Users/runner/miniforge3/conda-bld/biopython_1720014912120/work
chex==0.1.87
dm-haiku==0.0.13
dm-tree==0.1.8
immutabledict==4.2.1
jax==0.4.36
jaxlib==0.4.36
jmp==0.0.4
ml_collections==1.0.0
ml_dtypes==0.5.0
numpy @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_falv04vejw/croot/numpy_and_numpy_base_1730835597469/work/dist/numpy-2.1.3-cp312-cp312-macosx_11_0_arm64.whl#sha256=7164b5c0130ef176555189929b287c3ccb0546509c32dcb3f43ed4aaf45e2d24
opt_einsum==3.4.0
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
scipy==1.14.1
setuptools==73.0.1
six==1.17.0
tabulate==0.9.0
toolz==1.0.0
typing_extensions==4.12.2
tzdata==2024.2
wheel==0.45.1
